requirements.txt
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-kms==2.15.0
amqp==5.0.6
asgiref==3.3.4
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2018.1.18
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
crcmod==1.7
cryptography==3.4.7
distro-info===0.18ubuntu0.18.04.1
Django==3.1.3
django-filter==2.4.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-oss-storage==1.1.1
django-ratelimit==3.0.1
djangorestframework==3.12.2
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
importlib-metadata==4.6.1
incremental==16.10.1
jmespath==0.10.0
jsonfield==3.1.0
keyring==10.6.0
keyrings.alt==3.0
kombu==5.1.0
language-selector==0.1
mysqlclient==2.0.3
netifaces==0.10.4
oss2==2.14.0
PAM==0.4.2
prompt-toolkit==3.0.19
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.10.1
pygobject==3.26.1
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.4
python-dateutil==2.8.1
python-debian==0.1.32
python-memcached==1.59
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
redis==3.5.3
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
sqlparse==0.4.1
ssh-import-id==5.7
swapper==1.1.2.post1
systemd-python==234
Twisted==17.9.0
typing-extensions==3.10.0.0
ufw==0.36
urllib3==1.22
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.5.0
zope.interface==4.3.2
